pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "schwab-mcp"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0.0",
    "schwabdev>=3.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
schwab-mcp = "schwab_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/schwab_mcp"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "pytest-mock>=3.0.0"]
